Extract properties and items search logic
